rating

+++ database/migrations/2015_10_30_120454_create_article_ratings_table.php
@@ -17,7 +17,7 @@ class CreateArticleRatingsTable extends Migration {
 			$table->increments('id');
 			$table->integer('article_id')->unsigned();
 			$table->integer('rater_id')->unsigned();
-			$table->integer('rating')->default(0.0);
+			$table->double('rating')->default(0.0);
 			$table->timestamps();
 		});
 	}
